+++ Samples~/Assets/Tests/Runtime/Scripts/XAsset.Resource.Test.cs
@@ -4,9 +4,9 @@
 
 #if UNITY_INCLUDE_TESTS
 using NUnit.Framework;
+using System.Collections;
 using UnityEngine;
 using UnityEngine.TestTools;
-using System.Collections;
 using static EFramework.Asset.XAsset;
 
 [PrebuildSetup(typeof(TestXAssetBuild))]
@@ -15,6 +15,7 @@ public class TestXAssetResource
     [OneTimeSetUp]
     public void Init()
     {
+        Const.bBundleMode = true;
         Const.bundleMode = true;
         Bundle.Initialize();
     }
@@ -28,30 +29,52 @@ public class TestXAssetResource
         Bundle.Initialize();
     }
 
-    [TestCase(true)]
-    [TestCase(false)]
-    public void Load(bool bundleMode)
+    [TestCase(true, true)]
+    [TestCase(true, false)]
+    [TestCase(false, true)]
+    [TestCase(false, false)]
+    public void Load(bool bundleMode, bool referMode)
     {
         LogAssert.ignoreFailingMessages = true;
 
+        Const.bBundleMode = true;
         Const.bundleMode = bundleMode;
+
+        Const.bReferMode = true;
+        Const.referMode = referMode;
+
         // Arrange
-        var path = "Assets/Tests/Runtime/Resources/Bundle/Prefab/TestCube";
+        var assetPath = "Assets/Tests/Runtime/Resources/Bundle/Prefab/TestCube";
         var notExistPath = "Assets/Tests/Runtime/Resources/Bundle/Prefab/NotExist";
+        var bundleName = Const.GetName(assetPath);
 
-        // Act
-        var asset1 = Resource.Load(path, typeof(GameObject));
-
-        // Assert
+        // 非泛型加载
+        var asset1 = Resource.Load(assetPath, typeof(GameObject), obtain: true) as GameObject;
         Assert.IsNotNull(asset1, "加载的资产不应为空。");
         Assert.IsInstanceOf<GameObject>(asset1, "加载的资产应为 GameObject 类型。");
-        Assert.DoesNotThrow(() => Resource.Unload(path));
+        if (bundleMode)
+        {
+            if (referMode) Assert.IsNotNull(asset1.GetComponent<Resource.Refer>(), "引用模式下 GameObject 实例上的 Resource.Refer 对象不应当为空。");
+            // 使用 AssetBundle 模式加载时，Resource.Refer 组件会在源实例上保持，此处非 refer 模式下不作验证。
+            // else Assert.IsNull(asset1.GetComponent<Resource.Refer>(), "非引用模式下 GameObject 实例上的 Resource.Refer 对象应当为空。");
 
-        // 测试泛型加载
-        var asset2 = Resource.Load<GameObject>(path);
+            var bundleInfo = Bundle.Find(bundleName);
+            Assert.AreEqual(1, bundleInfo.Count, "obtain = true 时引用计数应当为 1。");
+
+            // 卸载
+            Assert.DoesNotThrow(() => Resource.Unload(assetPath));
+            Assert.AreEqual(0, bundleInfo.Count, "资源卸载后的引用计数应当仍为 0。");
+        }
+
+        // 泛型加载
+        var asset2 = Resource.Load<GameObject>(assetPath, obtain: false);
         Assert.IsNotNull(asset2, "加载的资产不应为空。");
         Assert.IsInstanceOf<GameObject>(asset2, "加载的资产应为 GameObject 类型。");
-        Assert.DoesNotThrow(() => Resource.Unload(path));
+        if (bundleMode)
+        {
+            var bundleInfo = Bundle.Find(bundleName);
+            Assert.AreEqual(0, bundleInfo.Count, "obtain = false 时引用计数应当为 0。");
+        }
 
         var notExistAsset = Resource.Load(notExistPath, typeof(GameObject));
         Assert.IsNull(notExistAsset, "加载的资产应为空。");
@@ -65,38 +88,64 @@ public class TestXAssetResource
         LogAssert.ignoreFailingMessages = true;
 
         bool[] bundleModes = { true, false };
+        bool[] referModes = { true, false };
+
         foreach (var bundleMode in bundleModes)
         {
+            Const.bBundleMode = false;
             Const.bundleMode = bundleMode;
-            // Arrange
-            var path = "Assets/Tests/Runtime/Resources/Bundle/Prefab/TestCube";
-            var notExistPath = "Assets/Tests/Runtime/Resources/Bundle/Prefab/NotExist";
 
-            // Act
-            var handler1 = Resource.LoadAsync(path, typeof(GameObject), (asset) =>
+            foreach (var referMode in referModes)
             {
-                // Assert
-                Assert.IsNotNull(asset, "加载的资产不应为空。");
-                Assert.IsInstanceOf<GameObject>(asset, "加载的资产应为 GameObject 类型。");
-                Assert.DoesNotThrow(() => Resource.Unload(path));
-            });
-            yield return handler1;
+                Const.bReferMode = false;
+                Const.referMode = referMode;
 
-            // 测试泛型加载
-            var handler2 = Resource.LoadAsync<GameObject>(path, (asset) =>
-            {
-                // Assert
-                Assert.IsNotNull(asset, "加载的资产不应为空。");
-                Assert.IsInstanceOf<GameObject>(asset, "加载的资产应为 GameObject 类型。");
-                Assert.DoesNotThrow(() => Resource.Unload(path));
-            });
-            yield return handler2;
+                // Arrange
+                var assetPath = "Assets/Tests/Runtime/Resources/Bundle/Prefab/TestCube";
+                var notExistPath = "Assets/Tests/Runtime/Resources/Bundle/Prefab/NotExist";
+                var bundleName = Const.GetName(assetPath);
 
-            var handler3 = Resource.LoadAsync(notExistPath, typeof(GameObject), (asset) =>
-            {
-                Assert.IsNull(asset, "加载的资产应为空。");
-            });
-            yield return handler3;
+                // Act
+                var handler1 = Resource.LoadAsync(assetPath, typeof(GameObject), origin =>
+                {
+                    var asset = origin as GameObject;
+                    Assert.IsNotNull(asset, "加载的资产不应为空。");
+                    Assert.IsInstanceOf<GameObject>(asset, "加载的资产应为 GameObject 类型。");
+                    if (bundleMode)
+                    {
+                        if (referMode) Assert.IsNotNull(asset.GetComponent<Resource.Refer>(), "引用模式下 GameObject 实例上的 Resource.Refer 对象不应当为空。");
+                        // 使用 AssetBundle 模式加载时，Resource.Refer 组件会在源实例上保持，此处非 refer 模式下不作验证。
+                        // else Assert.IsNull(asset.GetComponent<Resource.Refer>(), "非引用模式下 GameObject 实例上的 Resource.Refer 对象应当为空。");
+
+                        var bundleInfo = Bundle.Find(bundleName);
+                        Assert.AreEqual(1, bundleInfo.Count, "obtain = true 时引用计数应当为 1。");
+
+                        // 卸载
+                        Assert.DoesNotThrow(() => Resource.Unload(assetPath));
+                        Assert.AreEqual(0, bundleInfo.Count, "资源卸载后的引用计数应当仍为 0。");
+                    }
+                }, obtain: true);
+                yield return handler1;
+
+                // 测试泛型加载
+                var handler2 = Resource.LoadAsync<GameObject>(assetPath, asset =>
+                {
+                    Assert.IsNotNull(asset, "加载的资产不应为空。");
+                    Assert.IsInstanceOf<GameObject>(asset, "加载的资产应为 GameObject 类型。");
+                    if (bundleMode)
+                    {
+                        var bundleInfo = Bundle.Find(bundleName);
+                        Assert.AreEqual(0, bundleInfo.Count, "obtain = false 时引用计数应当为 0。");
+                    }
+                }, obtain: false);
+                yield return handler2;
+
+                var handler3 = Resource.LoadAsync(notExistPath, typeof(GameObject), asset =>
+                {
+                    Assert.IsNull(asset, "加载的资产应为空。");
+                });
+                yield return handler3;
+            }
         }
 
         LogAssert.ignoreFailingMessages = false;